requirements.txt
django==1.8.14
djangorestframework
djangorestframework-jwt
